api/requirements.txt
Flask==2.3.3
ytmusicapi==1.3.2
python-dotenv==1.0.0
yt-dlp==2023.9.24
requests==2.31.0
numpy==1.24.3
scikit-learn==1.3.0
flask-cors==4.0.0
Werkzeug==2.3.7